agent-e-ai-analysis/requirements.txt
# === 에이전트 E — AI 3각도 분석 엔진 ===
# Python 3.11+

# 웹 프레임워크
fastapi==0.115.6
uvicorn[standard]==0.34.0

# AI 클라이언트
openai==1.58.1
google-generativeai==0.8.3
anthropic==0.39.0

# xAI Grok은 OpenAI 호환 API 사용 (openai 패키지로 처리)

# Supabase
supabase==2.11.0

# HTTP 비동기 클라이언트
aiohttp==3.11.11
httpx==0.28.1

# 환경변수
python-dotenv==1.0.1

# 유틸리티
pydantic==2.10.4
pydantic-settings==2.7.0

# 테스트 (개발용)
pytest==8.3.4
pytest-asyncio==0.24.0
